+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Texture Replacer: Replacement textures not cleared when the selected material is changed.
+
 ## [1.0.0] - 2023-11-08
 
 Initial release. Migrated from [Material Replacer](https://github.com/kurotu/MaterialReplacer).